Add support for input secrets
